Really fixed it this time.

+++ StarRotator.py
@@ -303,8 +303,10 @@ class StarRotator(object):
         from matplotlib.patches import Circle
         import shutil
         import os
+        import os.path
+        if os.path.isdir('anim/') == True:
+            shutil.rmtree('anim/')#First delete the contents of the anim folder.
         os.mkdir('anim/')
-        shutil.rmtree('anim/')#First delete the contents of the anim folder.
         minflux = min(self.lightcurve)
         F=self.stellar_spectrum
         for i in range(self.Nexp):
